Simplify app to single record button UI
- Remove Next.js boilerplate and footer
- Consolidate recording functionality to main page
- Replace alerts with inline success/error icons
- Fix Vercel Blob upload error by using Buffer
- Clean up unused styles and components
