--- app/voice-recorder/voice-recorder.module.css
@@ -1,186 +1,0 @@
-.container {
-  min-height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  padding: 20px;
-}
-
-.recorder {
-  background: white;
-  border-radius: 20px;
-  padding: 40px;
-  box-shadow: 0 20px 60px rgba(0, 0, 0, 0.3);
-  max-width: 500px;
-  width: 100%;
-}
-
-.title {
-  text-align: center;
-  color: #333;
-  margin-bottom: 30px;
-  font-size: 2rem;
-  font-weight: 600;
-}
-
-.visualizer {
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  border-radius: 15px;
-  padding: 30px;
-  margin-bottom: 30px;
-  text-align: center;
-}
-
-.recordingIndicator {
-  color: white;
-  font-size: 1.2rem;
-  margin-bottom: 10px;
-  font-weight: 500;
-  transition: all 0.3s ease;
-}
-
-.recordingIndicator.active {
-  animation: pulse 1.5s infinite;
-}
-
-@keyframes pulse {
-  0% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0.6;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-
-.timer {
-  color: white;
-  font-size: 2.5rem;
-  font-weight: bold;
-  font-family: 'Courier New', monospace;
-}
-
-.controls {
-  display: flex;
-  gap: 15px;
-  justify-content: center;
-  margin-bottom: 20px;
-}
-
-.recordButton,
-.pauseButton,
-.stopButton {
-  padding: 12px 24px;
-  border: none;
-  border-radius: 50px;
-  font-size: 1rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.recordButton {
-  background: #ef4444;
-  color: white;
-}
-
-.recordButton:hover {
-  background: #dc2626;
-  transform: translateY(-2px);
-  box-shadow: 0 5px 15px rgba(239, 68, 68, 0.3);
-}
-
-.pauseButton {
-  background: #f59e0b;
-  color: white;
-}
-
-.pauseButton:hover {
-  background: #d97706;
-  transform: translateY(-2px);
-  box-shadow: 0 5px 15px rgba(245, 158, 11, 0.3);
-}
-
-.stopButton {
-  background: #6b7280;
-  color: white;
-}
-
-.stopButton:hover {
-  background: #4b5563;
-  transform: translateY(-2px);
-  box-shadow: 0 5px 15px rgba(107, 114, 128, 0.3);
-}
-
-.recordIcon,
-.pauseIcon,
-.stopIcon {
-  font-size: 1.2rem;
-}
-
-.playback {
-  margin-top: 20px;
-  padding: 20px;
-  background: #f3f4f6;
-  border-radius: 10px;
-}
-
-.playback h3 {
-  color: #333;
-  margin-bottom: 10px;
-  font-size: 1.1rem;
-}
-
-.audioPlayer {
-  width: 100%;
-}
-
-.uploadStatus {
-  text-align: center;
-  color: #667eea;
-  font-weight: 500;
-  margin-top: 20px;
-  padding: 15px;
-  background: #f0f4ff;
-  border-radius: 10px;
-  animation: fadeIn 0.3s ease;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-    transform: translateY(-10px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-@media (max-width: 640px) {
-  .recorder {
-    padding: 30px 20px;
-  }
-  
-  .title {
-    font-size: 1.5rem;
-  }
-  
-  .controls {
-    flex-direction: column;
-    width: 100%;
-  }
-  
-  .recordButton,
-  .pauseButton,
-  .stopButton {
-    width: 100%;
-    justify-content: center;
-  }
-}
